update docstring and move part of model builder in the timingmodel
